+++ application/src/test/java/core/application/users/service/DibServiceImplTest.java
@@ -1,7 +1,7 @@
 package core.application.users.service;
 
 import core.application.movies.models.entities.CachedMovieEntity;
-import core.application.movies.repositories.CachedMovieRepository;
+import core.application.movies.repositories.movie.CachedMovieRepository;
 import core.application.users.models.dto.DibRespDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
